ecommerce/requirements.txt
# Scrapy project requirements
Scrapy==2.11.0
scrapy-playwright==0.0.34
playwright==1.40.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
itemadapter==0.8.0
Twisted[http2]==23.10.0
